requirements.txt
bs4
haystack-ai
markdownify
tiktoken
streamlit
jinja2
diskcache